requirements.txt
streamlit
openai
python-dotenv
tenacity
rich
